AdventureGame-TB-FINAL
Final versions of all Java files for the text-based version of the game. This includes Junit tests for TicTacToeTest, PoliceStationTest, LocationATest, LocationBTest, LocationCTest, LocationDTest. The inheritance can be found mostly in the parent class PoliceStation where the child classes are the various locations which share a toString method.

The text-based version and GUI version of the games share the same logic. However, they are separate interfaces (one for text, and one for GUI). And the Junit tests are for the logic classes.